// website/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as admin from "../admin.js";
import type * as auth from "../auth.js";
import type * as cron from "../cron.js";
import type * as entitlements from "../entitlements.js";
import type * as http from "../http.js";
import type * as maintenance from "../maintenance.js";
import type * as serviceAuth from "../serviceAuth.js";
import type * as stripe from "../stripe.js";
import type * as stripeStore from "../stripeStore.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  auth: typeof auth;
  cron: typeof cron;
  entitlements: typeof entitlements;
  http: typeof http;
  maintenance: typeof maintenance;
  serviceAuth: typeof serviceAuth;
  stripe: typeof stripe;
  stripeStore: typeof stripeStore;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};